+++ sssom/validators.py
@@ -8,12 +8,10 @@ from linkml.validators.jsonschemavalidator import JsonSchemaDataValidator
 from linkml.validators.sparqlvalidator import SparqlDataValidator  # noqa: F401
 from sssom_schema import MappingSet
 
+from sssom.constants import SCHEMA_YAML, SchemaValidationType
 from sssom.context import add_built_in_prefixes_to_prefix_map
 from sssom.parsers import to_mapping_set_document
-from sssom.schema import SCHEMA_YAML
 from sssom.util import MappingSetDataFrame, get_all_prefixes
-
-from .constants import SchemaValidationType
 
 
 def validate(
